+++ packages/orq-rc/RELEASES.md
@@ -4739,3 +4739,13 @@ Based on:
 - [typescript v4.0.0-rc.2] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.2 - packages/orq-rc
+
+## 2025-11-03 14:54:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.646.0 (2.735.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.3] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.3 - packages/orq-rc